Moved FindJWT and ClaimsFromJWT jwtmanager
